requirements.txt
aniso8601==8.0.0
click==7.1.1
Flask==1.1.2
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.8
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
PyExifTool==0.1.1
pymongo==3.4.0
pytz==2019.3
six==1.14.0
Werkzeug==1.0.1
